Тесты Захар (#43)
* vacancy preview

* mock

* fix server mistakes, fix get resume allow, fix tests

* vacancyActivity handlers tests

* vacancyActivity handlers tests

* uncomment + fix current time

* more tests

* fix images naming

* fix

* delete session middleware from getting resume

* fix images

* fix apply

* tests for usecase mail

* add test for sendCode handler

* all tests for mail service

* tests for vacancy rep

* add more testCases for sign-in & sign-up handlers

* more test case

* woops..

* more tests for sign-in

* testsZakhar

* tests for usecase auth_micro

* few fixes...

* fix shlyapa

* few fix

* fix tests

* fix applies(kostilno)

* fix vet

* tests for handler auth_microservice

* fix run go test in ci

* def

* tests

+++ internal/usecases/impl/vacancy_activity.go
@@ -29,11 +29,12 @@ func (vas *VacancyActivityService) ApplyForVacancy(email string, vacancyId uint,
 	}
 
 	input.UserAccountId = user.ID
-	input.VacancyId = uint(vacancyId)
+	input.VacancyId = vacancyId
 	return vas.vacancyActivityRep.ApplyForVacancy(input)
 }
 
 func (vas *VacancyActivityService) GetAllUserApplies(userId uint) ([]*models.VacancyActivity, error) {
+
 	return vas.vacancyActivityRep.GetAllUserApplies(userId)
 }
